+++ acceleration-everest-03-dotnet/Infrastructure.Data/Mappings/PortifolioMapping.cs
@@ -33,7 +33,7 @@ public class PortifolioMapping : IEntityTypeConfiguration<Portfolio>
             .HasColumnName("AccounBalance");
 
         builder.Property(portfolio => portfolio.CustomerId)
-            .HasColumnType("INT")
+            .HasColumnType("BIGINT")
             .IsRequired()
             .HasColumnName("CustomerId");
